+++ castor/lib/core/src/main/scala/io/iohk/atala/castor/core/model/did/w3c/W3CModelHelper.scala
@@ -1,42 +1,47 @@
 package io.iohk.atala.castor.core.model.did.w3c
 
 import io.iohk.atala.castor.core.model.did.{
-  CanonicalPrismDID,
   DIDData,
   DIDMetadata,
+  PrismDID,
   PublicKey,
   PublicKeyData,
   Service,
   VerificationRelationship
 }
+import io.iohk.atala.shared.models.HexStrings.*
 
 object W3CModelHelper extends W3CModelHelper
 
 private[castor] trait W3CModelHelper {
 
   extension (didMetadata: DIDMetadata) {
-    def toW3C: DIDDocumentMetadataRepr = DIDDocumentMetadataRepr(deactivated = didMetadata.deactivated)
+    def toW3C(did: PrismDID): DIDDocumentMetadataRepr = DIDDocumentMetadataRepr(
+      deactivated = didMetadata.deactivated,
+      canonicalId = did.asCanonical.toString,
+      versionId = HexString.fromByteArray(didMetadata.lastOperationHash.toArray).toString
+    )
   }
 
   extension (didData: DIDData) {
-    def toW3C: DIDDocumentRepr = {
-      val keyWithPurpose = didData.publicKeys.map(k => k.purpose -> k.toW3C(didData.id, didData.id))
+    def toW3C(did: PrismDID): DIDDocumentRepr = {
+      val keyWithPurpose = didData.publicKeys.map(k => k.purpose -> k.toW3C(did, did))
       DIDDocumentRepr(
-        id = didData.id.toString,
-        controller = didData.id.toString,
+        id = did.toString,
+        controller = did.toString,
         verificationMethod = Nil,
         authentication = keyWithPurpose.collect { case (VerificationRelationship.Authentication, k) => k },
         assertionMethod = keyWithPurpose.collect { case (VerificationRelationship.AssertionMethod, k) => k },
         keyAgreement = keyWithPurpose.collect { case (VerificationRelationship.KeyAgreement, k) => k },
         capabilityInvocation = keyWithPurpose.collect { case (VerificationRelationship.CapabilityInvocation, k) => k },
         capabilityDelegation = keyWithPurpose.collect { case (VerificationRelationship.CapabilityDelegation, k) => k },
-        service = didData.services.map(_.toW3C(didData.id))
+        service = didData.services.map(_.toW3C(did))
       )
     }
   }
 
   extension (service: Service) {
-    def toW3C(did: CanonicalPrismDID): ServiceRepr = ServiceRepr(
+    def toW3C(did: PrismDID): ServiceRepr = ServiceRepr(
       id = s"${did.toString}#${service.id}",
       `type` = service.`type`.name,
       serviceEndpoint = service.serviceEndpoint.map(_.toString)
@@ -44,7 +49,7 @@ private[castor] trait W3CModelHelper {
   }
 
   extension (publicKey: PublicKey) {
-    def toW3C(did: CanonicalPrismDID, controller: CanonicalPrismDID): PublicKeyRepr = PublicKeyRepr(
+    def toW3C(did: PrismDID, controller: PrismDID): PublicKeyRepr = PublicKeyRepr(
       id = s"${did.toString}#${publicKey.id}",
       `type` = "EcdsaSecp256k1VerificationKey2019",
       controller = controller.toString,