+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/http/HttpRoutes.scala
@@ -5,7 +5,6 @@ import akka.http.scaladsl.server.Directives.*
 import io.iohk.atala.agent.openapi.api.{
   PresentProofApi,
   DIDApi,
-  DIDAuthenticationApi,
   DIDRegistrarApi,
   IssueCredentialsProtocolApi,
   ConnectionsManagementApi
@@ -16,19 +15,16 @@ import akka.http.scaladsl.server.Route
 object HttpRoutes {
 
   def routes: URIO[
-    DIDApi & DIDAuthenticationApi & DIDRegistrarApi & IssueCredentialsProtocolApi & ConnectionsManagementApi &
-      PresentProofApi,
+    DIDApi & DIDRegistrarApi & IssueCredentialsProtocolApi & ConnectionsManagementApi & PresentProofApi,
     Route
   ] =
     for {
       didApi <- ZIO.service[DIDApi]
-      didAuthApi <- ZIO.service[DIDAuthenticationApi]
       disRegistrarApi <- ZIO.service[DIDRegistrarApi]
       issueCredentialsProtocolApi <- ZIO.service[IssueCredentialsProtocolApi]
       connectionsManagementApi <- ZIO.service[ConnectionsManagementApi]
       presentProofApi <- ZIO.service[PresentProofApi]
     } yield didApi.route ~
-      didAuthApi.route ~
       disRegistrarApi.route ~
       issueCredentialsProtocolApi.route ~
       connectionsManagementApi.route ~